requirements.txt
voluptuous
attrs==19.1.0
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
jsonschema==3.0.1
pyrsistent==0.14.11
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
